Tried to solve a bug in src/lockelements.c
When locking a selection of elements, it is not easy to unselect the
selection since those elements and their pins/pads/elementlines/elementarcs
are now locked ;-)
This appears to be a bug in pcb.
A possible workaround would be to clear the selected flag of the
locked element at the same instance the lock flag was set.
This needs some further tweaking.

GUI is updated when leaving the command dialog.